BezierSwitchControlState 파일 분리
